minor fixes & add interfaces for graph & implement dfs path recursively
