+++ UnitTests/Algorithms/QuickSortTests.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Algorithms.Sorting;
 
 namespace UnitTests.Algorithms
